+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Remove unused ChannelPosition API.
 - Implement Endpoint and Format Enumeration for macos.
 - Implement format handling for macos `build_voice` method.